chore: implement panic catcher as the helper
